Daily view finished and Event Edit Done
Daily view now expands an event to the bottom of the layout if it ends
beyond the displayed day. Event Edit page now has all required features
(but no safety checks for events beginning after they end). That will be
an easy fix later if there is time. Can be resolved by dynamically
setting the min/max of the other pickers.